The first page is handled as LEAF page.
The first page is not only LEAF but also ROOT.
Previously, we handle that page as ROOT, that is, it has a abbrev key.
But that's not make sense.
This changes so that that page is handled as LEAF, and setting
abbrev key is occurred when inserting into parent page.
When first splitting, we get abbrev key from real leading value, and
then set to parent page (first root page).