Use tuplesort batch memory for randomAccess sorts
The tuplesort.c batch memory optimization was previously only used for
the final on-the-fly merge phase of an external sort, if any.  It makes
sense that batch memory cannot be used until the final merge phase,
since the number of tapes is fixed by then.  If batch memory was used
for every merge pass, the non-uniform state of each tape would imply
significant additional complexity and overhead, which seems very
unappealing.  However, there is no good reason to restrict the
optimization to final merges that happen to be performed on-the-fly
(because final output needs to be on a materialized tape).  Any final
merge should be comparable to a final on-the-fly merge in most of the
ways that matter, so we now apply the batch memory optimization for all
final merges.

Note that with this change, reloading no longer necessarily occurs for
all active tapes in the event of running out of preloaded tuples for
just one tape.  This does have a downside, as logtape.c prefers that
access patterns are characterized by alternately reading and writing as
much as possible at once.

While this optimization is beneficial to a variety of tuplesort.c
callers, the need for parallel worker merging in a later commit will
make it particularly compelling.  This parallel merging will have much
in common with the final merge needed for randomAccess cases.