Add force_btree_randomaccess GUC for testing
The tuplesort.c parallel sort infrastructure is intended as generic
infrastructure, usable in principle through any existing tuplesort
interface (in other words, any existing class of caller tuple can be
sorted in parallel, including for example heap tuplesort tuples).
Callers need only find a way to pass tuples from within worker
processes, and coordinate, per the tuplesort.h parallel sort contract.

Since the only current caller of the parallel sort interface is CREATE
INDEX, there is no preexisting way that a parallel sort could hit
various tuplesort.c codepaths that are needed for randomAccess cases
that happen to be performed in parallel.  The new
force_btree_randomaccess GUC offers a way to artificially force a
"randomAccess = true" argument for CREATE INDEX, to facilitate testing.
This may also be used to get a sense of the performance characteristics
of parallel sorting in isolation from index creation.

This patch is not proposed for inclusion in PostgreSQL; it is only for
testing during the review of the parallel CREATE INDEX patch.