Introduce GUC_NO_RESET flag.
Previously, GUCs having GUC_NO_RESET_ALL flag were ignored when the
RESET ALL command but were reset when the RESET command. So the
transaction-property GUCs such as transaction_isolation could be reset
after starting the transaction, which led to an assertion failure as
per reported in bug #17385.

This change adds a new flag GUC_NO_RESET and marks
transaction-property GUCs and seed GUC_NO_RESET. GUCs having
GUC_NO_RESET are forbidden to be reset by the RESET command as well as
be saved and restored.

Although it's uncommon that the transaction-property GUCs are reset
after BEGIN TRANSACTION, there may be functions that have any
optimization that depends on transaction_read_only as follow:

CREATE FUNCTION foo() RETURNS ...
SET transaction_read_only = 1;

This change prohibits this case too.

Since this change would break the compatibility of functions we don't
back-patch it.