Preparatory refactoring for decoupling kind from size class
Rename the current kind info array to refer to size classes, but
keep all the contents the same.

Add a fanout member to all nodes which stores the max capacity of
the node. This is currently set with the same hardcoded value as
in the kind info array.

In passing, remove outdated reference to node16 in the regression
test.