Make all node kinds variable sized
Add one size class 1, 15, and 61, for each node kind 4, 32, and 128,
respectively. The inner and leaf node size with the new size classes
are 24/24, 160/160, and 752/768, respectively.

For example in size class 15, when a 16th element is to be inserted,
allocte a larger area and memcpy the entire old node to it.

This technique allows us to limit the node kinds to 4, which
1. limits the number of cases in switch statements
2. allows a possible future optimization to encode the node kind
in a pointer tag