Add TIDStore, to store sets of TIDs (ItemPointerData) efficiently.
The TIDStore is designed to store large sets of TIDs efficiently, and
is backed by the radix tree.

The TIDStore is not used for anything yet, aside from the test code,
but a future commit will integrate the TIDStore with lazy vacuum,
reducing lazy vacuum memory usage and lifting the 1GB limit on its
size, by storing the list of dead TIDs more efficiently.

This includes a unit test module, in src/test/modules/test_tidstore.

WIP: Uses something similar to a PagetableEntry
but, for now, without a header. Also doesn't store
a "1" offset in the zero bit position. Needed for future
emdedding in a pointer type.