Use TIDStore for storing dead tuple TID during lazy vacuum.
Previously, we used an array of ItemPointerData to store dead tuple
TIDs, which is not space efficient and slow to lookup. Also, we had
the 1GB limit on its size.

This changes to use TIDStore for this purpose. Since the TIDStore,
backed by the radix tree, incrementaly allocates the memory, we get
rid of the 1GB limit.

Also, since we are no longer able to exactly estimate the maximum
number of TIDs can be stored based on the amount of memory. It also
changes to the column names max_dead_tuples and num_dead_tuples and to
show the progress information in bytes.

Furthermore, since the TIDStore use the radix tree internally, the
minimum amount of memory required by TIDStore is 1MB, which is the
inital DSA segment size. Due to that, this change increase the minimum
maintenance_work_mem from 1MB to 2MB.