Move some fields to LVVacCounters struct.
This commit is a preliminary work for paralle heap vacuuming. Some
fields in LVRelStates will be updated not only by the leader process
but also by parallel vacuum workers. These fields are now moved to the
new struct LVVacCounters, and LVRelStates has a pointer to it.

Reviewed-by:
Discussion: https://postgr.es/m/